Windows Service support
Update the project to generate two distributions. A `unix` distribution
that uses `jsvc` and a `windows` distribution using `winsw`.

The distribution is configurable via Maven properties which is useful as
Windows use the name of the exe as an identifier.

Right now, the JMX endpoint used to shutdown the app gracefully is
unsecured but the `StopService` class can be easily extended to
accommodate those requirements.