Update groupId to allow releases to maven central
Will be releasing my own versions of these artifacts to Maven repos, rather
than providing pull requests to upstream. Upstream is an excellent sample
of how to run Spring Boot as a Windows Service. These releases extend that
to provide that ability with an as close to a drop-in dependency as
possible.